refactor(#3265): extract standalone Proxy dispatch into object-runtime-proxy.ts
Behaviour-preserving god-file split (subtask of #3182). Moves the standalone
Proxy meta-object dispatch subsystem — ensureProxyRuntime, fillProxyDispatch,
and the 12 PROXY_CALL_* driver-name consts (1465 LOC, verbatim) — out of the
11,609-LOC object-runtime.ts into a new same-dir sibling module
src/codegen/object-runtime-proxy.ts.

Pure relocation, zero logic changes:
- new module exports ensureProxyRuntime (added the export keyword — the only
  edit to moved code) + the already-exported fillProxyDispatch; imports its
  deps at identical relative paths (same dir) plus ObjectRuntimeTypes/
  reserveApplyClosure back from ./object-runtime.js (tolerated function-level
  cycle).
- object-runtime.ts imports ensureProxyRuntime back (still called by
  ensureObjectRuntime) and re-exports fillProxyDispatch so index.ts's
  `from "./object-runtime.js"` importer keeps resolving.

Acceptance: npx tsc --noEmit → 0; prove-emit-identity check → IDENTICAL (39/39
across gc/standalone/wasi). Ratchets green; only check:coercion-sites tripped on
the relocated __is_truthy sites (0→3), granted via coercion-sites-allow in the
issue frontmatter (#3131, per-issue not whole-tree baseline). Smoke test
tests/issue-3265.test.ts (get/has/absent-trap/gc-lane) 4/4.


✓
